move context class
