+++ lib/zakuro/context/context.rb
