+++ spec/zakuro/calculation/option/dropped_date/location_spec.rb
@@ -3,7 +3,7 @@
 require File.expand_path('../../../../../lib/zakuro/calculation/option/dropped_date/location',
                          __dir__)
 
-require File.expand_path('../../../../../lib/zakuro/version/context',
+require File.expand_path('../../../../../lib/zakuro/context/context',
                          __dir__)
 
 require File.expand_path('../../../../../lib/zakuro/version/senmyou/cycle/solar_term',
