+++ spec/zakuro/version/senmyou/monthly/lunar_phase_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require File.expand_path('../../../../../' \
-  'lib/zakuro/version/context',
+  'lib/zakuro/context/context',
                          __dir__)
 
 require File.expand_path('../../../../../' \
